movies_admin/src/requirements.txt
django==4.2.5
flake8==6.1.0
django-split-settings==1.2.0
psycopg2-binary==2.9.9
black==24.2.0
uwsgi==2.0.20